CLI Structure, Initial Documentation
In this commit, I played around with the CLI structure for a while to
find abstractions that would allow me to divide up responsibilities
clearly, define asynchronous execution cleanly, and could scale to the
size of projcet this is likely to become.

The fundamental challenge with Ecology V1 was that I was looking for the
fastest path to get to a working command line tool, and in doing so I
made the code really hard to maintain. My hope for this rewrite is to
add structure and abstraction that enables the CLI to scale over time to
(hopefully) feature parity with Ecology + beyond.